Adding `with` support for consumers.
Creating a consumer using a `with` statement:

```python
  with Consumer(...) as c:
      c.subscribe(...)
      ...
```

...ensures that the consumer will be automatically closed when the
block exits. There is no longer any need for a `finally: close()`
block.

This change makes it easier for beginners and experts alike to use
consumers correctly and cleanly. The fact that it's less to type is a
nice bonus.

`with` has been a part of Python since v2.5 and is the recommended way
to manage resource lifecycles.