Add missing dependency on googleapis-common-protos when using protobufs (#1881)
When using the Schema Registry with Protobufs,
`confluent_kafka/schema_registry/protobuf.py` imports from the `google.type`
package, which is part of the googleapis-common-protos library, but the library
is not specified as a dependency.

This adds that library as a dependency to make sure it is installed, so users
won't hit an issue trying to import to Protobuf schema registry code in case
they don't have googleapis-common-protos installed.