Don't leave NewTopic partially-initialized on error
If `NewTopic` was initialized with a non-dict `config`, then the
constructor raised an exception but left the object in a
partially-initialized state with `replica_assignment` and `config`
elements that hadn't had their reference counts increased, leading to
mysterious segfaults in the test suite.  Be more careful to only set
these elements once we're ready to do so.

Fixes: #2146